add: members as a list
