--- src/constants/members.json
@@ -1,22 +1,0 @@
-{
-   "members":[
-    {
-        "id": 2223,
-        "name": "Kjetil Indrehus",
-        "study": "Dataingeniør 2.året",
-        "role": "Nettside Utvikler",
-        "attribute": ["It", "podcast"],
-        "description": "Liker å kode på fritida og lage mye gøy"
-
-    },
-    {
-        "id": 333,
-        "name": "Morten Jentoftsen",
-        "study": "Elektroingeniør - 3.året",
-        "role": "Promo Ansvarlig",
-        "attribute": ["promo", "styret"],
-        "description": "Liker å kode på fritida og lage mye gøy"
-
-    }
-   ]
-}
