--- src/pages/andreas.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const andreas = () => {
-  return (
-    <div>andreas</div>
-  )
-}
-
-export default andreas;
